requirements.txt

langchain_core
langchain-community
beautifulsoup4
langchain_groq
duckduckgo-search
wikipedia
arxiv 
langchain-ollama 
pypdf
chromadb 
streamlit
google-search-results 
faiss-cpu
selenium
beautifulsoup4
lxml 
html5lib
streamlit-authenticator
PYYaml
authlib
langchain-chroma